add plotly version of visual, better interaction

+++ common/visual/__init__.py
@@ -34,7 +34,7 @@ def get_colors(color='dark', to_int=True, opa=None, to_np=False):
 
     Args:
         color: color name
-        to_int: If true, change range from (0, 1)-float64 to (0, 255)-int8
+        to_int: If true, change range from (0, 1)-float64 to (0, 255)-int64
         opa: opacity value in (0, 1)
         to_np: if True, get numpy array
 
@@ -47,7 +47,7 @@ def get_colors(color='dark', to_int=True, opa=None, to_np=False):
     if opa is not None:
         col.append(opa)
     if to_np:
-        col = np.array(col, dtype=np.int8 if to_int else np.float64)
+        col = np.array(col, dtype=np.int64 if to_int else np.float64)
 
     return col
